requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
numpy==1.24.4
outcome==1.3.0.post0
pandas==2.0.3
psycopg2-binary==2.9.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
